# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiabetesInput {
  pregnancies: number;
  glucose: number;
  bloodPressure: number;
  skinThickness: number;
  insulin: number;
  bmi: number;
  diabetesPedigreeFunction: number;
  age: number;
}

export interface HeartInput {
  age: number;
  sex: number;
  chestPainType: number;
  restingBP: number;
  cholesterol: number;
  fastingBS: number;
  restingECG: number;
  maxHR: number;
  exerciseAngina: number;
  oldpeak: number;
  stSlope: number;
}

export interface ParkinsonsInput {
  mdvpFo: number;
  mdvpFhi: number;
  mdvpFlo: number;
  mdvpJitter: number;
  mdvpShimmer: number;
  nhr: number;
  hnr: number;
  rpde: number;
  dfa: number;
  spread1: number;
  spread2: number;
  d2: number;
  ppe: number;
}

export type PredictionResultPrediction =
  (typeof PredictionResultPrediction)[keyof typeof PredictionResultPrediction];

export const PredictionResultPrediction = {
  positive: "positive",
  negative: "negative",
} as const;

export type PredictionResultRiskLevel =
  (typeof PredictionResultRiskLevel)[keyof typeof PredictionResultRiskLevel];

export const PredictionResultRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface PredictionResult {
  prediction: PredictionResultPrediction;
  probability: number;
  riskLevel: PredictionResultRiskLevel;
  message: string;
  details: string;
}

export type PredictionRecordDiseaseType =
  (typeof PredictionRecordDiseaseType)[keyof typeof PredictionRecordDiseaseType];

export const PredictionRecordDiseaseType = {
  diabetes: "diabetes",
  heart: "heart",
  parkinsons: "parkinsons",
} as const;

export type PredictionRecordPrediction =
  (typeof PredictionRecordPrediction)[keyof typeof PredictionRecordPrediction];

export const PredictionRecordPrediction = {
  positive: "positive",
  negative: "negative",
} as const;

export type PredictionRecordRiskLevel =
  (typeof PredictionRecordRiskLevel)[keyof typeof PredictionRecordRiskLevel];

export const PredictionRecordRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface PredictionRecord {
  id: string;
  diseaseType: PredictionRecordDiseaseType;
  prediction: PredictionRecordPrediction;
  probability: number;
  riskLevel: PredictionRecordRiskLevel;
  message: string;
  createdAt: string;
}

export interface ChatSession {
  id: string;
  title: string;
  /** Last message preview */
  preview: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  bot: "bot",
} as const;

export interface ChatMessage {
  id: string;
  sessionId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatSessionWithMessages {
  id: string;
  title: string;
  messages: ChatMessage[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateChatSessionInput {
  title: string;
}

export interface SendMessageInput {
  content: string;
}

export interface DeleteResult {
  success: boolean;
}