# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HealthPredict AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: predictions
    description: Disease prediction operations
  - name: chat
    description: Medical chatbot operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /predictions/diabetes:
    post:
      operationId: predictDiabetes
      tags: [predictions]
      summary: Predict diabetes risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiabetesInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"

  /predictions/heart:
    post:
      operationId: predictHeart
      tags: [predictions]
      summary: Predict heart disease risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"

  /predictions/parkinsons:
    post:
      operationId: predictParkinsons
      tags: [predictions]
      summary: Predict Parkinson's disease risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkinsonsInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"

  /predictions/history:
    get:
      operationId: getPredictionHistory
      tags: [predictions]
      summary: Get prediction history
      responses:
        "200":
          description: List of past predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredictionRecord"

  /chat/sessions:
    get:
      operationId: getChatSessions
      tags: [chat]
      summary: List all chat sessions
      responses:
        "200":
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"

  /chat/sessions/{sessionId}:
    get:
      operationId: getChatSession
      tags: [chat]
      summary: Get a chat session with messages
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionWithMessages"
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /chat/sessions/{sessionId}/messages:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI response
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DiabetesInput:
      type: object
      properties:
        pregnancies:
          type: number
        glucose:
          type: number
        bloodPressure:
          type: number
        skinThickness:
          type: number
        insulin:
          type: number
        bmi:
          type: number
        diabetesPedigreeFunction:
          type: number
        age:
          type: number
      required: [pregnancies, glucose, bloodPressure, skinThickness, insulin, bmi, diabetesPedigreeFunction, age]

    HeartInput:
      type: object
      properties:
        age:
          type: number
        sex:
          type: number
        chestPainType:
          type: number
        restingBP:
          type: number
        cholesterol:
          type: number
        fastingBS:
          type: number
        restingECG:
          type: number
        maxHR:
          type: number
        exerciseAngina:
          type: number
        oldpeak:
          type: number
        stSlope:
          type: number
      required: [age, sex, chestPainType, restingBP, cholesterol, fastingBS, restingECG, maxHR, exerciseAngina, oldpeak, stSlope]

    ParkinsonsInput:
      type: object
      properties:
        mdvpFo:
          type: number
        mdvpFhi:
          type: number
        mdvpFlo:
          type: number
        mdvpJitter:
          type: number
        mdvpShimmer:
          type: number
        nhr:
          type: number
        hnr:
          type: number
        rpde:
          type: number
        dfa:
          type: number
        spread1:
          type: number
        spread2:
          type: number
        d2:
          type: number
        ppe:
          type: number
      required: [mdvpFo, mdvpFhi, mdvpFlo, mdvpJitter, mdvpShimmer, nhr, hnr, rpde, dfa, spread1, spread2, d2, ppe]

    PredictionResult:
      type: object
      properties:
        prediction:
          type: string
          enum: [positive, negative]
        probability:
          type: number
        riskLevel:
          type: string
          enum: [low, moderate, high]
        message:
          type: string
        details:
          type: string
      required: [prediction, probability, riskLevel, message, details]

    PredictionRecord:
      type: object
      properties:
        id:
          type: string
        diseaseType:
          type: string
          enum: [diabetes, heart, parkinsons]
        prediction:
          type: string
          enum: [positive, negative]
        probability:
          type: number
        riskLevel:
          type: string
          enum: [low, moderate, high]
        message:
          type: string
        createdAt:
          type: string
      required: [id, diseaseType, prediction, probability, riskLevel, message, createdAt]

    ChatSession:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        preview:
          type: string
          description: Last message preview
        messageCount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, preview, messageCount, createdAt, updatedAt]

    ChatSessionWithMessages:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, messages, createdAt, updatedAt]

    ChatMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum: [user, bot]
        content:
          type: string
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]

    CreateChatSessionInput:
      type: object
      properties:
        title:
          type: string
      required: [title]

    SendMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]