# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from "./chatMessageRole";

export interface ChatMessage {
  id: string;
  sessionId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}