# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSession {
  id: string;
  title: string;
  /** Last message preview */
  preview: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}