# lib/api-zod/src/generated/types/createChatSessionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChatSessionInput {
  title: string;
}