# lib/api-zod/src/generated/types/parkinsonsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParkinsonsInput {
  mdvpFo: number;
  mdvpFhi: number;
  mdvpFlo: number;
  mdvpJitter: number;
  mdvpShimmer: number;
  nhr: number;
  hnr: number;
  rpde: number;
  dfa: number;
  spread1: number;
  spread2: number;
  d2: number;
  ppe: number;
}