# lib/api-zod/src/generated/types/predictionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionRecordDiseaseType } from "./predictionRecordDiseaseType";
import type { PredictionRecordPrediction } from "./predictionRecordPrediction";
import type { PredictionRecordRiskLevel } from "./predictionRecordRiskLevel";

export interface PredictionRecord {
  id: string;
  diseaseType: PredictionRecordDiseaseType;
  prediction: PredictionRecordPrediction;
  probability: number;
  riskLevel: PredictionRecordRiskLevel;
  message: string;
  createdAt: string;
}