# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthPredict AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionResultPrediction } from "./predictionResultPrediction";
import type { PredictionResultRiskLevel } from "./predictionResultRiskLevel";

export interface PredictionResult {
  prediction: PredictionResultPrediction;
  probability: number;
  riskLevel: PredictionResultRiskLevel;
  message: string;
  details: string;
}